Renamed facade.py to gcp.py and added Facade class
