+++ ScoutSuite/providers/gcp/facade/gcp.py
@@ -1,0 +1,14 @@
+from ScoutSuite.providers.gcp.facade.facade import Facade
+from ScoutSuite.providers.gcp.facade.stackdriverlogging import StackdriverLoggingFacade
+from ScoutSuite.providers.gcp.facade.utils import GCPFacadeUtils
+
+class GCPFacade(Facade):
+    def __init__(self):
+        super(GCPFacade, self).__init__('cloudresourcemanager', 'v1')
+        self.stackdriverlogging = StackdriverLoggingFacade()
+
+    async def get_projects(self):
+        resourcemanager_client = self._get_client()
+        request = resourcemanager_client.projects().list() 
+        projects_group = resourcemanager_client.projects()
+        return await GCPFacadeUtils.get_all('projects', request, projects_group)
